use lastest version of tblprint lib
and require tblprint on all hook files
