+++ mods/Locked Down Regions/Server_AdvanceTurn.lua
@@ -1,3 +1,4 @@
+require('tblprint');
 require('_util');
 require('version');
