+++ mods/Locked Down Regions/Server_Created.lua
@@ -1,3 +1,5 @@
+require('tblprint');
+
 function Server_Created(game, settings)
 	local publicGD = Mod.PublicGameData;
